### 2019-06-25 #1527 @whedon
[PRE REVIEW]: Agents.jl: agent-based modeling framework in Julia
**Submitting author:** @kavir1698 (<a href="http://orcid.org/0000-0003-0895-1495">Ali Rezaee Vahdati</a>)
**Repository:** <a href="https://github.com/kavir1698/Agents.jl" target ="_blank">https://github.com/kavir1698/Agents.jl</a>
**Version:** v1.1.2
