### 2019-06-30 #1527 @kyleniemeyer
Hi @jedbrown, based on editor load and area I'm going to assign this one to you. Thanks!
