### 2019-07-02 #1527 @jedbrown
@kavir1698 Thanks for your submission. I see a few things that should be corrected before starting the review.
1. There are some syntactic issues with the references in the paper, e.g., `@[key]` should be `[@key]`. There is no need to cite your own repository; it already appears in the margin decoration.
2. The paper is very short and would benefit from some greater explanation of features, perhaps a performance comparison or brief example to help a prospective user determine whether Agents.jl will meet their needs.
3. It looks like Travis-CI is used only to build documentation. Can you make it run the automated test suite? In any case, please document how to run the test suite.
4. Please add CONTRIBUTING.md and address [Community Guidelines](https://joss.readthedocs.io/en/latest/review_checklist.html).
5. This software claims performance as a feature so the docs should define a comparison and enable users to replicate it. In most cases, the paper should summarize such results.
