### 2019-07-03 c9b3d2a
Update packages, modify the paper, etc
