+++ CONTRIBUTING.md
@@ -1,0 +1,8 @@
+# Contributing to Agents.jl
+
+Thanks for taking the time to contribute.
+
+Any contribution to Agents is welcome, especifically via the following ways:
+
+  * Modifying the code with a pull request
+  * Reporting bugs and suggestions in the issues section of the project's Github.
