+++ src/Agents.jl
@@ -13,6 +13,7 @@ using CSV
 using ColorTypes
 using Compose
 using GraphPlot
+import Cairo, Fontconfig
 import Base.iterate
 import Base.length
 # using PerceptualColourMaps
