+++ .travis.yml
@@ -4,6 +4,7 @@ os:
 
 julia:
   - 1.0
+  - 1.1
 notifications:
   email: false
 git:
@@ -23,3 +24,6 @@ jobs:
                                                  Pkg.instantiate()'
           - julia --project=docs/ docs/make.jl
         after_success: skip
+
+script:
+- julia --color=yes -e 'using Pkg; Pkg.build(); Pkg.test()'
