+++ README.md
@@ -1,6 +1,6 @@
 # Agents: agent-based modeling framework in Julia
 
-[![](https://img.shields.io/badge/Agents.jl-v1.2-blue.svg)](https://github.com/kavir1698/Agents.jl) 
+[![](https://img.shields.io/badge/Agents.jl-v1.1.2-blue.svg)](https://github.com/kavir1698/Agents.jl) 
 [![](https://img.shields.io/badge/docs-stable-blue.svg)](https://kavir1698.github.io/Agents.jl/stable)
 [![](https://img.shields.io/badge/docs-latest-blue.svg)](https://kavir1698.github.io/Agents.jl/dev)
 [![Build Status](https://travis-ci.org/kavir1698/Agents.jl.svg?branch=master)](https://travis-ci.org/kavir1698/Agents.jl)
