### 2019-07-05 c9f101b
Bug fix and minor changes
