+++ src/data_collector.jl
@@ -184,7 +184,7 @@ end
 Combines columns of the data that contain the same type of info from different steps of the model into one column using an aggregator, e.g. mean. You should either supply all column names that contain the same type of data, or one name (as a string) that precedes a number in different columns, e.g. "pos_"{some number}.
 """
 function combine_columns!(data::DataFrame, column_base_name::String, aggregators::AbstractVector)
-  column_names = vcat([column_base_name], [column_base_name*"_"*string(i) for i in 1:size(data)[2]])
+  column_names = vcat([column_base_name], [column_base_name*"_"*string(i) for i in 1:size(data, 2)])
   datanames = [string(i) for i in names(data)]
   final_names = Array{Symbol}(undef, 0)
   for cn in column_names
