+++ src/Agents.jl
@@ -12,10 +12,10 @@ import StatsBase
 using DataVoyager
 using CSV
 using ColorTypes
-using Compose
 using GraphPlot
 using Cairo
 using Fontconfig
+using Compose
 import Base.iterate
 import Base.length
 include("agents_component.jl")
