### 2019-07-22 9f2bc22
Agents.jl vs Mesa benchmark, update paper
Plus minor improvements in data collector
