### 2019-07-23 #1527 @Datseris
Hi, yes I'd be happy to review @jedbrown . This seems very interesting for me, and I didn't know if its existence before now.
