### 2019-07-28 ffa4a4b
Refactored schelling.jl in examples
