+++ examples/schelling.jl
@@ -2,73 +2,110 @@
 
 * There are agents on a grid
 * Agents are of two different kinds. They can represent different groups.
-* Only one agent can be at a node on the grid. Each node represents a house, and only one a person/family can live in a house.
+* Only one agent can be at a node on the grid.
+  * Each node represents a house, and only one a person/family can live in a house.
 * Agents are happy even if a majority of their neighbors are of a different kind.
 * If they are unhappy, they move to a random empty node until they are happy.
 
-This model shows that even a slight preference for being around neighbors of the same kind leads to segregated neighborhoods.
+This model shows that even a slight preference for being around neighbors of the same kind leads to segregated
+neighborhoods.
+
 =#
 
 using Agents
 
-# Create agent, model, and grid types
-mutable struct SchellingAgent{T<:Integer} <: AbstractAgent # An agent
-# object should always be a subtype of AbstractAgent
+# Create the agent, model, and grid types.
+"AbstractModel type for the Schelling Model"
+mutable struct SchellingModel{T<:Integer, Y<:AbstractArray, Z<:AbstractSpace} <: AbstractModel 
+  # Object should always be a subtype of AbstractModel.
+  "A field of the model for a space object, always a subtype of AbstractSpace."
+  space::Z 
+  "A list of agents."
+  agents::Y
+  "A field for the scheduler function."
+  scheduler::Function
+  "The minimum number of neighbors for agent to be happy."
+  min_to_be_happy::T
+end
+
+"AbstractAgent type for the Schelling Agent"
+mutable struct SchellingAgent{T<:Integer} <: AbstractAgent # Object should always be a subtype of AbstractAgent
+  "The identifier number of the agent."
   id::T
+  "The x, y location of the agent."
   pos::Tuple{T, T}
-  mood::Bool # true is happy and false is unhappy
+  "Whether or not the agent is happy with cell (true is 'happy' and false is 'unhappy')."
+  mood::Bool
+  "The group of the agent, determines mood as it interacts with neighbors."
   group::T
 end
 
-mutable struct SchellingModel{T<:Integer, Y<:AbstractArray, Z<:AbstractSpace} <: AbstractModel  # A model
-	# object should always be a subtype of AbstractModel
-	space::Z  # A space object, which is a field
-	# of the model object is always subtype of AbstractSpace
-	agents::Y  # a list of agents
-	scheduler::Function
-	min_to_be_happy::T  # minimum number of neighbors 
-	#to be of the same kind so that they are happy
-end
-
+"The space of the experiment."
 mutable struct MyGrid{T<:Integer, Y<:AbstractArray} <: AbstractSpace
+  "Dimensions of the grid."
   dimensions::Tuple{T, T}
+  "The space type."
   space::SimpleGraph
-  agent_positions::Y  # an array of arrays for each grid node
+  "An array of arrays for each grid node."
+  agent_positions::Y  
 end
 
-# instantiate the model
+"Function to instantiate the model."
 function instantiate_model(;numagents=320, griddims=(20, 20), min_to_be_happy=3)
+  # Create an empty Array of Arrays.
   agent_positions = [Int64[] for i in 1:gridsize(griddims)]
-  mygrid = MyGrid(griddims, grid(griddims, false, true), agent_positions)  # create a 2D grid with nodes have a max of 8 neighbors
+  # Use MyGrid to create a grid from griddims and agent_positions using the grid function.
+  # Create a 2D grid with nodes have a max of 8 neighbors.
+  mygrid = MyGrid(griddims, grid(griddims, false, true), agent_positions)
[…]
