### 2019-07-28 e4c6be7
Copied code from schelling.jl to schelling.md
