+++ docs/src/schelling.md
@@ -26,11 +26,22 @@ The `scheduler` field accepts a function that defines the order with which agent
 ```julia
 using Agents
 
-mutable struct SchellingModel <: AbstractModel  # A model object should always be a subtype of AbstractModel
- space::AbstractSpace  # A space object, which is a field of the model object is always subtype of AbstractSpace
- agents::Array{AbstractAgent}  # a list of agents
- scheduler::Function
- min_to_be_happy::Integer  # minimum number of neighbors to be of the same kind so that they are happy
+"""
+AbstractModel type for the Schelling Model
+
+Object should always be a subtype of AbstractModel.
+"""
+mutable struct SchellingModel{T<:Integer, Y<:AbstractArray,
+                              Z<:AbstractSpace} <: AbstractModel 
+
+  "A field of the model for a space object, always a subtype of AbstractSpace."
+  space::Z
+  "A list of agents."
+  agents::Y
+  "A field for the scheduler function."
+  scheduler::Function
+  "The minimum number of neighbors for agent to be happy."
+  min_to_be_happy::T
 end
 ```
 
@@ -41,11 +52,25 @@ It is best to make any model parameter a field of the model object. We add the m
 Next, we define an agent object. Agent objects are subtypes of `AbstractAgent` and should always have the following fields: `id` which stores agent IDs as integers, and `pos` to store each agent's position. Agent positions can be tuple of integers as coordinates of nodes of a grid (1D, 2D or 3D). Positions can also be integers only, referring to the number of a node in an irregular network.
 
 ```julia
-mutable struct SchellingAgent <: AbstractAgent # An agent object should always be a subtype of AbstractAgent
- id::Integer
- pos::Tuple{Integer, Integer}
- mood::Bool # true is happy and false is unhappy
- group::Integer
+"""
+AbstractAgent type for the Schelling Agent
+
+Object should always be a subtype of AbstractAgent.
+"""
+mutable struct SchellingAgent{T<:Integer} <: AbstractAgent
+  "The identifier number of the agent."
+  id::T
+  "The x, y location of the agent."
+  pos::Tuple{T, T}
+  """
+  Whether or not the agent is happy with cell.
+
+  Where true is "happy" and false is "unhappy"
+
+  """
+  mood::Bool
+  "The group of the agent, determines mood as it interacts with neighbors."
+  group::T
 end
 ```
 
@@ -58,10 +83,14 @@ Finally, we define a space object. The space object is always a subtype of `Abst
 The second field of the space object is the `dimensions` of the grid or network. Lastly, every space object should have an `agent_positions` field. This field is an array of arrays for each node of the network. Each inner array will record the ID of the agents on that position. Agents.jl keeps the position of agents in two places. One in each agent's object and one in the `agent_positions`.
 
 ```julia
-mutable struct MyGrid <: AbstractSpace # A space object should always be a subtype of AbstractSpace
- dimensions::Tuple{Integer, Integer}
- space
- agent_positions::Array  # an array of arrays for each grid node
+"The space of the experiment."
+mutable struct MyGrid{T<:Integer, Y<:AbstractArray} <: AbstractSpace
+  "Dimensions of the grid."
+  dimensions::Tuple{T, T}
+  "The space type."
+  space::SimpleGraph
+  "An array of arrays for each grid node."
+  agent_positions::Y  
 end
 ```
 
@@ -70,28 +99,46 @@ end
 Now that we have defined the basic objects, we should instantiate the model. We put the model instantiation in a function so that it will be easy to recreate the model and change its parameters.
 
 ```julia
-function instantiate_model(; numagents=320, griddims=(20, 20), min_to_be_happy=3)
[…]
