### 2019-07-28 29c30c1
Added descriptive paragraph to top of schelling.md
