### 2019-07-30 #1527 @Datseris
@jedbrown I am currently on vacation and I have some free spare time. Is it possible for me to start the review process now, or in the next couple of days? Because after my vacation I will not be able to do the review (too busy with PhD defense).
