### 2019-07-30 JuliaDynamics/Agents.jl#15 @Datseris
Would you like to join JuliaDynamics?
Hey there,

let me start with an important declaration: **The discussion here will in no way affect my review at [JOSS](https://github.com/openjournals/joss-reviews/issues/1527).**

Agents.jl is a Julia package that revolves around agent-based-modeling and cellular automata. I unfortunately was not aware about this until I was given the opportunity to review it... JuliaDynamics is a GitHub organization about dynamical systems and nonlinear dynamics (and also recently for assisting scientific projects). It contains several packages already, one being the award-winning **DynamicalSystems.jl**. What is notably missing from JuliaDynamics are systems with discrete space and time (i.e. cellular automata).

https://juliadynamics.github.io/JuliaDynamics/

Agents.jl seems to fit in JuliaDynamics from a scientific perspective. In addition, its documentation clearly matches the standards of JuliaDynamics, providing examples, tutorials and scientific motivation on the use of ABM. At JuliaDynamics we try to be a coherent unit ("organization") that can spread word and interest for dynamical systems. It is also very good for users to have packages of similar scientific orientation to organizations. It makes them easier to find and more trustworthy

I would like to invite you to join us at JuliaDynamics. What we will gain is that your excellent work will be part of our collective effort for creating high quality libraries for dynamical systems. What you'll gain by joining:

* A logo for Agents.jl
* Help in maintaining the repo
* Pull Requests review
* Documentation improvements (if necessary)
* Be part of an already recognized (and highly active) GitHub org.
* More exposure: easier to find it if part of JuliaDynamics, and also showcased in our website. I think you could benefit from this.

Let me know what you think! In a similar way that I approached you, I also approached [RecurrenceAnalysis.jl](https://github.com/JuliaDynamics/RecurrenceAnalysis.jl/issues/8).

*please notice that for conflict-of-interest purposes, this can only progress from my side after the review of JOSS is over*
