### 2019-07-30 kavir1698/Agents.jl#15 @Datseris
(I forgot to mention the obvious technicality: should you chose to join, the moment you do you will become part of a team that has owner-level access to this repo)
