### 2019-07-30 kavir1698/Agents.jl#15 @kavir1698
Hi,
Thank you for the invitation. I am happy to join the JuliaDynamics organization, as it will help introduce, maintain, and develop Agents.jl.
