### 2019-07-30 #1527 @Datseris
@jedbrown I want to be sure there isn't any conflict of interest here. @kavir1698 has agreed to join us in JuliaDynamics, a GitHub organization that I am part of. In this process I will do some contributions to the repo under review here. But, since the rules of JOSS state that I can't review a repo that I contribute to, I will do this _after_ the current review process is done.

is this okay?
