### 2019-07-30 kavir1698/Agents.jl#15 @def-mycroft
I've recently started investing some time into making contributions to Agents.jl. I'm very interested in using ABM techniques in economics research (I'm working on a graduate program in applied math right now in the US and plan to continue studies in economics). The Agents project provides such a useful framework for ABM work.

I definitely hope to continue making contributions to the project in the future.

The JuliaDynamics org and Agents.jl looks like a great fit!
