### 2019-07-30 #1527 @jedbrown
@Datseris Yes, I can start the review now. (I've been working to track down another reviewer, but can add them later since you're ready to get started.) Thanks for disclosing this potential CoI -- I don't consider plans to collaborate to be problematic for the review.
