### 2019-07-30 #1611 @whedon
[REVIEW]: Agents.jl: agent-based modeling framework in Julia
**Submitting author:** @kavir1698 (<a href="http://orcid.org/0000-0003-0895-1495">Ali Rezaee Vahdati</a>)
**Repository:** <a href="https://github.com/kavir1698/Agents.jl" target ="_blank">https://github.com/kavir1698/Agents.jl</a>
**Version:** v1.1.8

@Datseris and @mozhgan-kch, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
