### 2019-07-30 #1611 @jedbrown
@Datseris :wave: Welcome, and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the [Agents.jl repository](https://github.com/kavir1698/Agents.jl)). I'll be watching this thread if you have any questions.
