### 2019-07-30 #1611 @Datseris
Here are some initial comments for the review.

1. The currently installed version for the package manager is `1.1.0`. This does not match the submitted version of `1.1.2`, but I don't know how much of an impact this has.
