+++ test/runtests.jl
@@ -2,5 +2,7 @@ using Test
 using Agents
 
 include("forest_fire_defs.jl")
+include("schelling_defs.jl")
 include("space_test.jl")
 include("data_collector_test.jl")
+include("schelling_test.jl")
