### 2019-07-31 6df1590
Change version in README.md and add doi to a paper
