+++ paper/paper.bib
@@ -16,14 +16,16 @@ title = {{MASON: a multiagent simulation environment}},
 volume = {81},
 year = {2005}
 }
-@article{Masad2015,
+@inproceedings{Masad2015,
 author = {Masad, David and Kazil, Jacqueline},
-journal = {PROC. OF THE 14th PYTHON IN SCIENCE CONF},
 keywords = {cellular automata,complexity,modeling,simulation},
-number = {Scipy},
 pages = {53--60},
 title = {{Mesa: an agent-based modeling framework}},
-year = {2015}
+year = {2015},
+doi = {10.25080/majora-7b98e3ed-009},
+publisher = {{SciPy}},
+booktitle = {Proceedings of the 14th Python in Science Conference}
+}
 }
 @article{Wilensky1999,
 author = {Wilensky, Uri},
