### 2019-08-02 #1611 @Datseris
@jedbrown  @kavir1698 I think I have now finished the first round of review. The above checklist states what I feel should be taken care of before acceptance.
