### 2019-08-08 7306e7d
Fix the missing "green" color.
