### 2019-08-08 d39dba8
Update visualization.jl and the paper
Add argument to save figures in a custom location.
Update the docstrings of visualization functions.
Add the first para of the docs to the paper.
