+++ paper/paper.bib
@@ -1,3 +1,35 @@
+@article{Farmer2009,
+abstract = {The leaders of the world are flying the economy by the seat of their pants, say J. Doyne Farmer and Duncan Foley. There is, however, a better way to help guide financial policies.},
+author = {Farmer, J. Doyne and Foley, Duncan},
+doi = {10.1038/460685a},
+file = {:C$\backslash$:/Users/Ali/Dropbox/Papers/Farmer, Foley/Nature{\_}2009{\_}The economy needs agent-based modelling.pdf:pdf},
+issn = {0028-0836},
+journal = {Nature},
+mendeley-groups = {complexity},
+month = {aug},
+number = {7256},
+pages = {685--686},
+title = {{The economy needs agent-based modelling}},
+url = {http://www.nature.com/articles/460685a},
+volume = {460},
+year = {2009}
+}
+@article{grimm2006standard,
+author = {Grimm, Volker and Berger, Uta and Bastiansen, Finn and Eliassen, Sigrunn and Ginot, Vincent and Giske, Jarl and Goss-Custard, John and Grand, Tamara and Heinz, Simone K and Huse, Geir and Huth, Andreas and Jepsen, Jane U. and J{\o}rgensen, Christian and Mooij, Wolf M. and M{\"{u}}ller, Birgit and Pe'er, Guy and Piou, Cyril and Railsback, Steven F. and Robbins, Andrew M. and Robbins, Martha M. and Rossmanith, Eva and R{\"{u}}ger, Nadja and Strand, Espen and Souissi, Sami and Stillman, Richard A. and Vab{\o}, Rune and Visser, Ute and DeAngelis, Donald L.},
+doi = {10.1016/j.ecolmodel.2006.04.023},
+file = {:C$\backslash$:/Users/Ali/Dropbox/Papers/Grimm et al/Ecological modelling{\_}2006{\_}A standard protocol for describing individual-based and agent-based models.pdf:pdf},
+issn = {03043800},
+journal = {Ecological Modelling},
+mendeley-groups = {postdoct{\_}paper{\_}1},
+month = {sep},
+number = {1-2},
+pages = {115--126},
+publisher = {Elsevier},
+title = {{A standard protocol for describing individual-based and agent-based models}},
+url = {https://linkinghub.elsevier.com/retrieve/pii/S0304380006002043},
+volume = {198},
+year = {2006}
+}
 @article{Luke2005,
 abstract = {Keywords: Agent-based modeling, , - systems, computational science 1. Introduction ThebasicMASONdistributionprovidesonlythosecore tools common to most and visualization needs. Analysis.},
 archivePrefix = {arXiv},
