+++ paper/paper.md
@@ -17,7 +17,9 @@ bibliography: paper.bib
 
 # Summary
 
-Agent-based modeling involves designing a system of autonomous agents that interact based on a set of given rules. It is used for studying complex systems whose behavior cannot be easily identified using classical mathematical approaches. There are currently several agent-based modeling frameworks available (notable examples are NetLogo [@Wilensky1999], Repast [@North2013], MASON [@Luke2005], and Mesa [@Masad2015]), but not for the Julia language, a new language designed with the needs of scientific computing in mind, such as interactivity, speed, and expressiveness. I introduce the first of such frameworks in Julia: Agents.jl. An ABM framework in Julia provides advantages to the existing frameworks in other languages. First, unlike NetLogo, Agents.jl uses a general-purpose language rather than custom one, which reduces the learning curve and unifies the modeling and analysis language. Second, Julia is an easy to learn and easy to write language, unlike Java that is used for Repast and MASON, and provides a REPL (Read-Eval-Print-Loop) to build and analyze models interactively. Third, Julia is fast to run unlike Python used for Mesa (see Fig. 1). This is can be important in large agent-based models.
+Agent-based modeling involves designing a system of autonomous agents that interact based on a set of given rules. It is used for studying complex systems whose behavior cannot be easily identified using classical mathematical approaches. An agent-based (or individual-based) model is a computational simulation of autonomous agents that react to their environment (including other agents) given a predefined set of rules [@grimm2006standard]. ABM has gained wide usage in a variety of research disciplines. One reason for its popularity is that it allows relaxing many simplifying assumptions usually made by mathematical models. Relaxing such assumptions of a "perfect world" can change a model's behavior [@Farmer2009]. ABM is specifically an important tool for studying complex systems where a system's behavior cannot be predicted and has to be explored.
+
+There are currently several agent-based modeling frameworks available (notable examples are NetLogo [@Wilensky1999], Repast [@North2013], MASON [@Luke2005], and Mesa [@Masad2015]), but not for the Julia language, a new language designed with the needs of scientific computing in mind, such as interactivity, speed, and expressiveness. I introduce the first of such frameworks in Julia: Agents.jl. An ABM framework in Julia provides advantages to the existing frameworks in other languages. First, unlike NetLogo, Agents.jl uses a general-purpose language rather than custom one, which reduces the learning curve and unifies the modeling and analysis language. Second, Julia is an easy to learn and easy to write language, unlike Java that is used for Repast and MASON, and provides a REPL (Read-Eval-Print-Loop) to build and analyze models interactively. Third, Julia is fast to run unlike Python used for Mesa (see Fig. 1). This is can be important in large agent-based models.
 
 ![Speed comparison of a version of "forest fire" model in Agents.jl vs Mesa. The same implementation of the model in Agents.jl (originally taken from Mesa's example and then re-implemented in Agents.jl) shows more than 8x speed gain. See the docs for more details.](benchmark01.png)
 
