+++ src/Agents.jl
@@ -13,9 +13,8 @@ using DataVoyager
 using CSV
 using ColorTypes
 using GraphPlot
-using Fontconfig
+using Cairo, Fontconfig
 using Compose
-using Cairo
 import Base.iterate
 import Base.length
 include("agents_component.jl")
