### 2019-08-08 f869329
Refactor the docs tutorial
