+++ docs/src/schelling.md
@@ -1,244 +1,0 @@
-# Schelling's segregation model
-
-## Agents.jl's architecture
-
-Agents.jl is composed of components for building models, building and managing space structures, collecting data, running batch simulations, and data visualization.
-
-Agents.jl structures simulations in three components: a _model_ component that keeps all model-level variables and data, an _agent_ component that keeps all agent-level variables and data, and a _space_ component that keeps space-level data.
-
-For building any ABM, users have to define at least three objects (_model_, _agent_ and _space_) and one function (Fig. 1). Agents.jl's tools manage the rest of the path to producing data and visualizations (Fig. 1). We now demonstrate Agents.jl's architecture and features through building Schelling's segregation model.
-
-![Fig. 1. __Path from building a model to gaining information from the model using Agents.jl.__ The box in cyan is what the user has to provide and the boxes in green are what Agents.jl provides.](agentscomponents.png)
-
-We implement the following definition of Schelling's segregation model:
-
-* Agents are of two kind (0 or 1).
-* Each agent has eight neighbors (Moore neighborhoods).
-* If an agent is in the same group with at least three neighbors, then it is happy.
-* If an agent is unhappy, it keeps moving to new locations until it is happy.
-
-### Defining a model object
-
-Building models using Agents.jl, we always start by defining three basic objects: one for the model, one for the the agents and one for the space.
-
-A model object is a subtype of `AbstractModel`. Making the model a subtype of `AbstractModel` will make Agents.jl methods available to the model. It needs to have the following three fields: `scheduler`, `space`, and `agents`. We can add more fields if needed.
-
-The `scheduler` field accepts a function that defines the order with which agents will activate at each step. The function should accept the model object as its input and return a list of agent indices. Agents.jl provides three schedulers: `as_added` to activate agents as they have been added to the model, `random_activation` to activate agents randomly, and `partial_activation` to activate only a random fraction of agents at each step.
-
-```julia
-using Agents
-
-"""
-AbstractModel type for the Schelling Model
-
-Object should always be a subtype of AbstractModel.
-"""
-mutable struct SchellingModel{T<:Integer, Y<:AbstractArray,
-                              Z<:AbstractSpace} <: AbstractModel 
-
-  "A field of the model for a space object, always a subtype of AbstractSpace."
-  space::Z
-  "A list of agents."
-  agents::Y
-  "A field for the scheduler function."
-  scheduler::Function
-  "The minimum number of neighbors for agent to be happy."
-  min_to_be_happy::T
-end
-```
-
-It is best to make any model parameter a field of the model object. We add the minimum number of neighbors of the same kind for an agent to be happy as a field of the model (`min_to_be_happy`). 
-
-### Defining an agent object
-
-Next, we define an agent object. Agent objects are subtypes of `AbstractAgent` and should always have the following fields: `id` which stores agent IDs as integers, and `pos` to store each agent's position. Agent positions can be tuple of integers as coordinates of nodes of a grid (1D, 2D or 3D). Positions can also be integers only, referring to the number of a node in an irregular network.
-
-```julia
-"""
-AbstractAgent type for the Schelling Agent
-
-Object should always be a subtype of AbstractAgent.
-"""
-mutable struct SchellingAgent{T<:Integer} <: AbstractAgent
-  "The identifier number of the agent."
-  id::T
-  "The x, y location of the agent."
-  pos::Tuple{T, T}
-  """
-  Whether or not the agent is happy with cell.
-
-  Where true is "happy" and false is "unhappy"
-
-  """
-  mood::Bool
-  "The group of the agent, determines mood as it interacts with neighbors."
-  group::T
-end
-```
-
[…]
