+++ docs/src/tutorial.md
@@ -1,0 +1,256 @@
+# Agents.jl's architecture
+
+Agents.jl is composed of components for building models, building and managing space structures, collecting data, running batch simulations, and data visualization.
+
+Agents.jl structures simulations in three components: a _model_ component that keeps all model-level variables and data, an _agent_ component that keeps all agent-level variables and data, and a _space_ component that keeps space-level data.
+
+For building any ABM, users have to define at least three objects (_model_, _agent_ and _space_) and one function (Fig. 1). Agents.jl's tools manage the rest of the path to producing data and visualizations (Fig. 1).
+
+![Fig. 1. __Path from building a model to gaining information from the model using Agents.jl.__ The box in cyan is what the user has to provide and the boxes in green are what Agents.jl provides.](agentscomponents.png)
+
+A **model object** is a subtype of `AbstractModel`. Making the model a subtype of `AbstractModel` will make Agents.jl methods available to the model. It needs to have the following three fields: `scheduler`, `space`, and `agents`. More fields can be added if needed. It is best to make any model parameter a field of the model object.
+
+The **`scheduler` field** accepts a function that defines the order with which agents will activate at each step. The function should accept the model object as its input and return a list of agent indices. Agents.jl provides three schedulers: `as_added` to activate agents as they have been added to the model, `random_activation` to activate agents randomly, and `partial_activation` to activate only a random fraction of agents at each step.
+
+The **space object** is always a subtype of `AbstractSpace` and should have at least the following three fields.
+
+  * First, a **`space` field** which holds the spatial structure of the model. Agents.jl uses network structures from the [LightGraphs package](https://github.com/JuliaGraphs/LightGraphs.jl) to represent space.  It provides 1D, 2D and 3D grids. The grids may have periodic boundary conditions, meaning nodes on the left and right edges and top and bottom edges are connected to one another. Furthermore, the nodes on a grid may have von Neumann neighborhoods, i.e. only connect to their orthogonal neighbors, or Moore neighborhoods, i.e. connect to their orthogonal and diagonal neighbors. Users may also provide arbitrary networks as their model's spatial structure.
+  * Second, **`dimensions`** of the grid or network in a `Tuple`.
+  * Third, **`agent_positions`** field. This field is an array of arrays for each node of the network. Each inner array will record the ID of the agents on that position. Agents.jl keeps the position of agents in two places. One in each agent's object and one in the `agent_positions`.
+
+**Agent objects** are subtypes of `AbstractAgent` and should always have the following fields: `id` which stores agent IDs as integers, and `pos` to store each agent's position. Agent positions can be tuple of integers as coordinates of nodes of a grid (1D, 2D or 3D). Positions can also be integers only, referring to the number of a node in an irregular network.
+
+**_step_ function.** Any ABM model should have at least one and at most two step functions. An _agent step function_ is always required. Such an agent step function defines what happens to an agent when it activates. Sometimes we also need a function that changes all agents at once, or changes a model property. In such cases, we can also provide a _model step function_.
+
+An agent step function should only accept two arguments: first, an agent object, and second, a model object.
+
+The model step function should accept only one argument, that is the model object. It is possible to only have a _model step function_, in which case users have to use the built-in `dummystep` as the _agent step function_.
+
[…]
