### 2019-08-09 9b16776
Update grid docstring
