### 2019-08-09 #1611 @kavir1698
Thank you, @Datseris, for your review.

@jedbrown I have released a new version (v1.1.7) for the package. I do not know how to change the version in the paper, though.
