### 2019-08-09 kavir1698/Agents.jl#25 @Datseris
Use Tagbot to tag releases
So that a new version is updated in the Julia package ecosystem.

1. Install JuliaRegistrator: and https://github.com/JuliaRegistries/Registrator.jl/#registrator tagbot : https://github.com/apps/julia-tagbot
2. Update `Project.toml` to match the version you want to tag
3. Go into the commit and comment "@JuliaRegistrator register()"
4. Win!
