### 2019-08-09 kavir1698/Agents.jl#25 @JuliaRegistrator
Error while trying to register: **Register Failed**
@Datseris, it looks like you don't have collaborator status on this repository.
