### 2019-08-09 #1611 @Datseris
Please notice that for my review to conclude, a new version of the software should be released in the Julia package ecosystem. That will correspond to the version a user would obtain via normal installation, `pkg> add Agents`. Although this seems a technicality, it is nevertheless the way a Julia package is installed (unless one wants to mess with the master branch, where versions do not really matter).

I recommend @kavir1698 to leverage `JuliaRegistrator` as instructed here: https://github.com/kavir1698/Agents.jl/issues/25
