### 2019-08-10 kavir1698/Agents.jl#25 @kavir1698
OK, thanks for the tips. I have followed along. The pull request created by JuliaRegistrator does not get merged automatically and I don't see anywhere to confirm it. What am I missing here?
