### 2019-08-14 #1611 @kavir1698
Hi @Datseris, v1.1.7 is now released in the Julia package ecosystem.
