### 2019-08-14 #1611 @Datseris
Great! @jedbrown I conclude my review now. All points pass in my eyes.
