### 2019-08-16 #1611 @jedbrown
@mozhgan-kch :wave: Welcome, and thanks for agreeing to review! The comments from @whedon at the top of this thread outline the review process, which takes place in this thread (possibly with issues filed in the [Agents.jl repository](https://github.com/kavir1698/Agents.jl)). There is a review checklist for you above. I'll be watching this thread if you have any questions.
