+++ docs/src/boltzmann_example01.md
@@ -1,159 +1,222 @@
 # Boltzmann wealth distribution
 
-This is a simple agent-based model in economics. Despite its simplicity, it shows striking emergent wealth distribution. The first model that we will does not have spatial structure.
+The Boltzmann wealth distribution is a simple agent-based model (ABM) in economics. Despite its simplicity, the model shows striking emergent wealth distribution. This model is the first in this set of examples that does not necessarily have a spatial structure. At first, we will build a model that does not have a spatial structure, and then we will modify the model to include a spatial structure. 
 
-* We start with a number of agents all of which have one unit of wealth.
+The Boltzmann model is conceptually very simple:
+
+* We start with a number of agents, each of which have one unit of wealth.
 * At every step, agents give one unit of their money (if they have any) to a random agent.
 * We will see how wealth will be distributed after a few steps.
 
-The code of this tutorial is in the `examples/boltzmann_wealth_distribution.jl` file on the Github repository.
+The code referenced in this tutorial is available in the [`examples/boltzmann_wealth_distribution.jl`](https://github.com/kavir1698/Agents.jl/blob/master/examples/boltzmann_wealth_distribution.jl) and [`examples/boltzmann_wealth_distribution_with_grid.jl`](https://github.com/kavir1698/Agents.jl/blob/master/examples/boltzmann_wealth_distribution_with_grid.jl) files on the Github repository.
 
 ## Building the model
 
-Agents.jl structures simulations in three components. A _model_ component that keeps all model-level variables and data, an _agent_ component that keeps all agent-level variables and data, and _space_ component that keeps space-level data.
+Recall that Agents.jl structures simulations in three components: a _model_ component that keeps all model-level variables and data, an _agent_ component that keeps all agent-level variables and data, and a _space_ component that keeps space-level data. These components are provided in Agents.jl as abstract types, and this typing allows for tools from Agents.jl manage the rest of the path to producing data and visualizations.
 
-At the beginning of each building any model, define your types for each of these components. After that, you will have to initialize your model, and write one or two functions to change angets and/or the model at each step. This will be all before you can run your model and analyze its results.
-
-Now let's build our three types of this model:
-
-These types should be subtypes of the following abstract types:
+Subtyping the model components in the following way will allow all the built-in functions to work on your defined types:
 
 * The agent type should be a subtype of `AbstractAgent`.
 * The model type should be a subtype of `AbstractModel`.
 * The space type should be a subtype of `AbstractSpace`.
 
-This subtyping will allow all the built-in functions to work on your define types.
+At first in this example, we will not be using a spatial structure:
 
 ```julia
 using Agents
 
-# 1. define agent type
+"""
+Defines the agent type.
+
+The agent type must be a subtype of AbstractAgent.
+
+Commonly, an agent type will require a field for location value in the form
+`pos::Tuple{T, T}`. In the first part of this example we will not be using a spatial
+structure, therefore we will not define a field for position.
+
+"""
 mutable struct MyAgent{T<:Integer} <: AbstractAgent
+  "The identifier number of the agent."
   id::T
-  pos::Tuple{T, T}  # x,y coords
+  "The agent's wealth."
   wealth::T
 end
 ```
-The agent type has to have the `id` and the `pos` (for position) fields, but it can have any other fields that you desire. Here we add a `wealth` field that accepts integers. If your space is a grid, the position should accept a `Tuple{Integer, Integer}` representing x, y coordinates.
+
[…]
