### 2019-08-20 e2674b9
Updated agent_step! function in Boltzmann example.
This is for issue #27. Added new agent_step! code provided
by kavir1698, function forces agents to only give wealth
to someone on their own node.
