### 2019-08-20 3295789
Modified Boltzmann agent_step! to move agents.
Agents who trade wealth with only agents on their own
node must be moved at every step, otherwise they will
just trade wealth amongst themselves.
