+++ examples/boltzmann_wealth_distribution_with_grid.jl
@@ -105,6 +105,12 @@ function agent_step!(agent::AbstractAgent, model::AbstractModel)
     # Then decrement the current agent's wealth and increment the neighbor's wealth.
     agent.wealth -= 1
 	random_neighbor_agent.wealth += 1
+
+    # Now move the agent to a random node.
+    # If the agent weren't moved, agents would merely trade wealth 
+    # amongst themselves on the same node.
+    neighboring_nodes = node_neighbors(agent, model)
+    move_agent!(agent, rand(neighboring_nodes), model)
   end
 end
 
