### 2019-08-22 5580f03
Comply with latex LightGraph syntax for two fs
