### 2019-08-30 90e8ce5
Add some minor tests and functions
