### 2019-09-05 0ffd841
Make some small improvements
