### 2019-09-15 2ba39fe
Fix a bug in move_agent! fucntion
