### 2019-09-15 75b2f98
Fix a bug in data collection, update dependencies
