### 2019-09-20 #1611 @mozhgan-kch
Adding to the above, I would like to make below suggestions:

- [ ] It would be good to know if you have done any benchmarking to compare this with other frameworks such as Netlogo or Repast, since you mentioned these in your paper.
- [ ] It would be beneficial to read about the limitations of this tool/framework in the paper. If there is any.
- [ ] The last reference does not have a DOI.
- [ ] As part of the `README.md` , it would be good to have some note on how to contribute back to this code and repo or if anyone would like to request a feature, what would the procedure be.
- [ ] The motivation is not clear in the paper.
- [ ] There are various ways to study complex systems and one way is ABM. I suggest the author add a line or two on bottom-up vs top-down approach and why not ABM and not mathematical equations - see https://www.pnas.org/content/99/suppl_3/7280
- [ ] Lack of citations in _doc version 1.1.7_, section _Why we need agent-based modeling_.
- [ ] Please proofread the paper. (e.g: _This is can be important in large agent-based models_)
