### 2019-09-30 6e3303d
Update the paper
