+++ README.md
@@ -30,3 +30,15 @@ Install using the following command inside Julia:
 It is compatible with Julia 0.7+.
 
 For a tutorial, read the docs: [![](https://img.shields.io/badge/docs-stable-blue.svg)](https://kavir1698.github.io/Agents.jl/stable)
+
+
+## Contributions
+
+Any contribution to Agents.jl is welcome in the following ways:
+
+  * Modifying the code or documentation with a pull request.
+  * Reporting bugs and suggestions in the issues section of the project's Github.
+
+### Previewing Documentation Edits
+
+Modifications to the documentation can be previewed by building the documentation locally, which is made possible by a script located in docs/make.jl. The Documenter package is required and can be installed by running `import Pkg; Pkg.add("Documenter")` in a REPL session. Then the documentation can be built and previewed in build/ first by running `julia docs/make.jl` from a terminal.
