### 2019-09-30 #1611 @kavir1698
Dear @mozhgan-kch,

Thank you for your comments and review. Below are my responses to your comments.

1. I have provided a comparison of speed between Agents.jl and Mesa, since they are the most similar to each other. The advantage of Agents.jl over Repast and NetLogo is that Agents.jl is implemented in Julia, which is an easy-to-learn language that is more geared toward data analysis and scientific computing.

2. I have added a couple of sentences in the paper mentioning some limitations of Agents.jl: " In its current version, it does not provide tools to visualize simulations in real time. Moreover, a GUI would make the package even more accessible.".

3. The paper does not seem to have a DOI. I did not find any.

4. A note about contributions is in the CONTRIBUTING.md file. I have copied it at the end of the README.md.

5. I add the following sentences to the paper:
  "Julia language provides a combination of features that were historically mutually exclusive. Specifically, languages that were fast to write, such as Python, were slow to run. And languages that were fast to run, such as C/C++, were slow to write. The combination of these two features, and the expressive structure of the language, makes Julia a desirable choice for scientific purposes. Agent-based models can involve hundreds of thousands of agents, each of which performing certain computations at each time-step. Thus, having a modeling framework that makes writing models easier and results in fast code is an advantage."

6. I added the following paragraph:
  "ABM provides a bottom-up approach for studying complex systems, whereas analytical models have a top-down one [@Bonabeau2002]. The two approaches are complementary and they both can benefit from insights that the other approach contributes. Analytical models make many simplifying assumptions about the systems they study. This results in systems that are tractable and lead to clear conclusions. Agent-based models on the other hand, are more difficult to make sense of because they relax many assumptions of equation-based approaches. This is at the same time an advantage of agent-based models because it allows observing the effect of agent and environment heterogeneity and stochasticity, which can change a model's behavior [@Farmer2009]. ABM is specifically an important tool for studying complex systems where a system's behavior cannot be predicted and has to be explored."

7. I have now cited four papers on the first line of the section.

8. Thank you for mentioning the typo. I have checked the paper again.
