+++ paper/paper.bib
@@ -74,3 +74,16 @@ title = {{Complex adaptive systems modeling with Repast Simphony}},
 volume = {1},
 year = {2013}
 }
+@article{Bonabeau2002,
+author = {Bonabeau, E},
+doi = {10.1073/pnas.082080899},
+issn = {0027-8424},
+journal = {Proceedings of the National Academy of Sciences},
+month = {may},
+number = {Supplement 3},
+pages = {7280--7287},
+title = {{Agent-based modeling: Methods and techniques for simulating human systems}},
+url = {http://www.pnas.org/cgi/doi/10.1073/pnas.082080899},
+volume = {99},
+year = {2002}
+}
