### 2019-10-05 #1611 @jedbrown
Thanks for your reviews, @mozhgan-kch and @Datseris.

@kavir1698 I posted a PR above with some copy edits to the paper. I have just a couple comments before proceeding:
* The discussion of GUI features is awkward and I don't know what you intend. Is this suggesting that such a feature would be valuable future work? Could you please reword that bit?
* The *Features* section is currently a mix of phrases (with or without periods) and full sentences. It should be reworded for consistency.
