### 2019-10-05 #1611 @jedbrown
Could you please downcase "cellular automata" (not a proper noun in [existing publications](https://scholar.google.com/scholar?hl=en&q=cellular%20automata&as_sdt=1%2C6&as_sdtp=)), tag a release (annotated tag preferred), and archive your repository on Zenodo or similar. When finished, please report the DOI back here. Thanks.
