### 2019-10-07 7e0d373
Update the paper
