### 2019-10-07 #1611 @jedbrown
Yes, we ask to archive the post-review software and it should be tagged. It could be `v1.1.7.1` or `v1.1.8` as you see fit.
