### 2019-10-07 #1611 @kavir1698
Agents.jl v1.1.7 has all the changes that were suggested in the review process. Review number 2 did not require any changes to the code, but only to the paper. Shall I still make a new release?
