### 2019-10-07 #1611 @jedbrown
If changes were contained to the paper, then archiving v1.1.7 is okay, but if there are changes to project source or documentation, please tag a new release.
