### 2019-10-08 ed8e50f
Move to v1.1.8 for the paper
