### 2019-10-08 #1611 @kavir1698
Since there was a small change in the documentation, I tagged a new release. Here is the DOI from Zenodo: `doi.org/10.5281/zenodo.3477581`.
