### 2019-10-08 #1611 @jedbrown
doi.org is being slow. Can you please update the author list to match the paper (i.e., just you).
