### 2019-10-08 #1611 @labarba
I jus finished reading the paper—all good, except for the performance comparison in Figure 1, it would be nice if you provided the specs of the hardware used for the test!
