+++ paper/paper.md
@@ -23,7 +23,7 @@ Agent-based modeling provides a bottom-up approach for studying complex systems,
 
 There are currently several agent-based modeling frameworks available (notable examples are NetLogo [@Wilensky1999], Repast [@North2013], MASON [@Luke2005], and Mesa [@Masad2015]), but not for the Julia language, a new language designed with the needs of scientific computing in mind. Julia provides a combination of features that have historically been mutually exclusive. Specifically, languages that were fast to write, such as Python, were slow to run. And languages that were fast to run, such as C/C++, were slow to write. The combination of speed, expressiveness, and support for interactive computing makes Julia a desirable choice for scientific purposes. Agent-based models can involve hundreds of thousands of agents, each performing certain computations at each time step, thus speed is essential.  Agents.jl is the first high-performance agent-based modeling framework to be written in Julia, and offers key advantages relative to the existing frameworks in other languages. First, unlike NetLogo, Agents.jl uses a general-purpose language rather than custom one, which reduces the learning curve and unifies the modeling and analysis language. Second, Julia is an easy to learn and easy to write language, unlike Java that is used for Repast and MASON, and provides a REPL (Read-Eval-Print-Loop) to build and analyze models interactively. Third, Julia is fast to run, unlike Python, which is used for Mesa (see Figure 1). This can be important in large agent-based models.
 
-![Speed comparison of a version of the "forest fire" model in Agents.jl vs Mesa. The same implementation of the model in Agents.jl (originally taken from Mesa's example and then re-implemented in Agents.jl) shows more than 8x speed gain. See the documentation for more details.](benchmark01.png)
+![Speed comparison of a version of the "forest fire" model in Agents.jl vs Mesa. The same implementation of the model in Agents.jl (originally taken from Mesa's example and then re-implemented in Agents.jl) shows more than 8x speed gain. See the documentation for more details. The comparison was performed on a Windows machine with a i7-6500U CPU and 16 GB of RAM. ](benchmark01.png)
 
 Agents.jl allows users to only think about what needs to happen during each step of their model, and the rest will be managed by the framework. Future development can include building a GUI to help users with less programming knowledge, and implementing real-time visualization of simulations. The [documentation](https://kavir1698.github.io/Agents.jl/dev/) contains a tutorial and several example agent-based models to demonstrate the features and workflows supported by the package.
 
