### 2019-10-08 #1611 @kavir1698
I have added the following line to the figure caption:

"The comparison was performed on a Windows machine with a i7-6500U CPU and 16 GB of RAM."
