### 2019-10-08 #1611 @labarba
Congratulations, @kavir1698, your JOSS paper is now published! 🚀

Huge thanks to our editor: @jedbrown, and the reviewers: @Datseris, @mozhgan-kch — your contribution to JOSS is much appreciated 🙏
