### 2019-10-08 #1611 @labarba
@arfon — I just looked at the PDF, and it doesn't contain the last commit by the author. Is `@whedon` not able to catch a very recent commit?
