### 2019-10-08 #1611 @kyleniemeyer
@labarba @arfon could the issue be that the change was made after the initial `whedon accept` command, and then whedon used that version with the deposit?
