### 2019-10-08 #1611 @labarba
I don't think so. In another paper I just published, some copy edits I submitted via PR did not appear in the PDF doing a `@whedon accept` immediately after the merge, but after waiting a while, I ran it again, and they got caught.I have a feeling there's some caching delay or whatnot.
