### 2019-10-09 kavir1698/Agents.jl#15 @Datseris
Now that the JOSS paper is published, I think we can progress here! Let me summarize how I think a good plan moving forward is:

1. @kavir1698 moves the repo to JuliaDynamics, and I make a team that gives him owner level access.
2. I start working on #20
3. [Optional] @kavir1698 can attempt to tackle #24
4. Me and a friend of mine work on a logo for Agents.jl
5. I make an announcement at discourse about the Agents.jl

What do you think?
