### 2019-10-09 kavir1698/Agents.jl#15 @kavir1698
Pretty good. Let's move forward.
