### 2019-10-09 f08ce72
Update README.md to include the paper
