### 2019-10-09 #1611 @arfon
I just regenerated the PDF locally and updated the `joss-papers` repo with the new PDF.

This is showing up as fixed for me now but might take a few hours to show up as modified for some of you as there's caching in place for the PDFs.
